backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
google-generativeai
python-multipart
pypdf
duckduckgo-search
